+++ test/unit-test/test_mgmt_deploy_utils.py
@@ -21,7 +21,7 @@ import mock_data.mock_deployment_manifest_huge as mock_deployment_manifest_huge
 import mock_data.mock_manifests as mock_manifests
 import mock_data.mock_module_info_huge as mock_module_info_huge
 import pytest
-from moto import mock_sts
+from moto import mock_aws
 
 import seedfarmer.errors
 import seedfarmer.mgmt.deploy_utils as du
@@ -45,7 +45,7 @@ def aws_credentials():
 
 @pytest.fixture(scope="function")
 def sts_client(aws_credentials):
-    with mock_sts():
+    with mock_aws():
         yield boto3_client(service_name="sts", session=None)